fixes ghastly trampling.

+++ code/game/objects/items/lightstick.dm
@@ -10,49 +10,42 @@
 	l_color = "#47A3FF" //Blue
 	var/s_color = "blue"
 
-	Crossed(var/mob/living/O)
-		if(anchored && prob(20))
-			if(!istype(O,/mob/living/carbon/Xenomorph/Larva))
-				visible_message("<span class='danger'>[O] tramples the [src]!</span>")
-				playsound(src, 'sound/weapons/Genhit.ogg', 25, 1)
-				if(istype(O,/mob/living/carbon/Xenomorph))
-					if(prob(40))
-						qdel(src)
-					else
-						anchored = 0
-						icon_state = "lightstick_[s_color][anchored]"
-						SetLuminosity(0)
-						pixel_x = 0
-						pixel_y = 0
-				else
-					anchored = 0
-					icon_state = "lightstick_[s_color][anchored]"
-					SetLuminosity(0)
-					pixel_x = 0
-					pixel_y = 0
-
-	//Removing from turf
-	attack_hand(mob/user)
-		..()
-		if(!anchored)//If planted
+/obj/item/lightstick/Crossed(mob/living/O)
+	. = ..()
+	if(anchored && prob(20) && istype(O) && !isxenolarva(O))
+		visible_message("<span class='danger'>[O] tramples the [src]!</span>")
+		playsound(src, 'sound/weapons/Genhit.ogg', 25, 1)
+		if(isxeno(O) && prob(40))
+			qdel(src)
 			return
-
-		to_chat(user, "You start pulling out \the [src].")
-		if(!do_after(user,20, TRUE, src, BUSY_ICON_GENERIC))
-			return
-
-		anchored = 0
-		user.visible_message("[user.name] removes \the [src] from the ground.","You remove the [src] from the ground.")
+		anchored = FALSE
 		icon_state = "lightstick_[s_color][anchored]"
 		SetLuminosity(0)
 		pixel_x = 0
 		pixel_y = 0
-		playsound(user, 'sound/weapons/Genhit.ogg', 25, 1)
+
+	//Removing from turf
+/obj/item/lightstick/attack_hand(mob/user)
+	. = ..()
+	if(!anchored)//If planted
+		return
+
+	to_chat(user, "You start pulling out \the [src].")
+	if(!do_after(user,20, TRUE, src, BUSY_ICON_GENERIC))
+		return
+
+	anchored = FALSE
+	user.visible_message("[user.name] removes \the [src] from the ground.","You remove the [src] from the ground.")
+	icon_state = "lightstick_[s_color][anchored]"
+	SetLuminosity(0)
+	pixel_x = 0
+	pixel_y = 0
+	playsound(user, 'sound/weapons/Genhit.ogg', 25, 1)
 
 	//Remove lightsource
-	Destroy()
-		SetLuminosity(0)
-		. = ..()
+/obj/item/lightstick/Destroy()
+	SetLuminosity(0)
+	return ..()
 
 /obj/item/lightstick/anchored
 	icon_state = "lightstick_blue1"
